refactor: extract useTranscription and compose useDictation
