+++ src/lib/useDictation.ts
@@ -1,13 +1,9 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useCallback, useMemo, useRef, useState } from 'react';
 import { useToast } from '@/components/ui/toast';
-import { mapAiError } from '@/features/ai/aiErrors';
 import { useTextStructuring } from '@/features/ai/useTextStructuring';
 import { addItemsToBin } from '@/features/bins/useBins';
-import { apiFetch } from '@/lib/api';
-import type { RecordingHandle } from '@/lib/audioRecorder';
-import { startRecording } from '@/lib/audioRecorder';
-import { Events, notify } from '@/lib/eventBus';
 import { clientItemId } from '@/lib/itemQuantities';
+import { useTranscription } from '@/lib/useTranscription';
 import type { AiSuggestedItem, BinItem } from '@/types';
 
 export type DictationState =
@@ -17,10 +13,6 @@ export type DictationState =
   | 'previewing-transcript'
   | 'structuring'
   | 'preview';
-
-interface TranscribeResponse {
-  text: string;
-}
 
 interface UseDictationOptions {
   binId?: string;
@@ -35,21 +27,14 @@ export function useDictation(options: UseDictationOptions) {
   const { binId, binName, existingItems, locationId, onAdd } = options;
   const { showToast } = useToast();
 
-  // Stabilize existingItems to avoid rebuilding the callback chain on every render.
-  // The key is a content-based string so the memo only updates when items actually change.
   const existingItemsKey = existingItems.join('\0');
   // biome-ignore lint/correctness/useExhaustiveDependencies: intentional content-based stabilization
   const stableExistingItems = useMemo(() => existingItems, [existingItemsKey]);
 
-  const [state, setState] = useState<DictationState>('idle');
+  const [itemState, setItemState] = useState<'idle' | 'previewing-transcript' | 'structuring' | 'preview'>('idle');
   const [transcript, setTranscript] = useState<string | null>(null);
-  const [error, setError] = useState<string | null>(null);
-  const [duration, setDuration] = useState(0);
-
-  const recordingRef = useRef<RecordingHandle | null>(null);
   const autoAdvanceRef = useRef<ReturnType<typeof setTimeout> | null>(null);
   const cancelledRef = useRef(false);
-  const startRef = useRef<() => void>(() => {});
 
   const {
     structuredItems,
@@ -57,41 +42,9 @@ export function useDictation(options: UseDictationOptions) {
     clearStructured,
   } = useTextStructuring();
 
-  // Tick elapsed time during recording
-  const isRecording = state === 'recording';
-  useEffect(() => {
-    if (!isRecording) return;
-    setDuration(0);
-    const id = setInterval(() => setDuration((d) => d + 1), 1000);
-    return () => clearInterval(id);
-  }, [isRecording]);
-
-  // Cleanup on unmount
-  useEffect(() => {
-    return () => {
-      recordingRef.current?.cancel();
-      if (autoAdvanceRef.current) clearTimeout(autoAdvanceRef.current);
-    };
-  }, []);
-
-  const cancel = useCallback(() => {
-    cancelledRef.current = true;
-    recordingRef.current?.cancel();
-    recordingRef.current = null;
-    if (autoAdvanceRef.current) {
-      clearTimeout(autoAdvanceRef.current);
-      autoAdvanceRef.current = null;
-    }
-    setState('idle');
-    setTranscript(null);
-    setError(null);
-    setDuration(0);
-    clearStructured();
-  }, [clearStructured]);
-
   const advanceToStructuring = useCallback(
     (text: string) => {
-      setState('structuring');
+      setItemState('structuring');
       structure({
         text,
         mode: 'items',
@@ -100,111 +53,58 @@ export function useDictation(options: UseDictationOptions) {
       }).then((items) => {
         if (cancelledRef.current) return;
         if (items && items.length > 0) {
-          setState('preview');
+          setItemState('preview');
         } else {
           showToast({ message: 'No items found in transcription', variant: 'warning' });
-          setState('idle');
+          setItemState('idle');
           setTranscript(null);
           clearStructured();
         }
       }).catch(() => {
         if (cancelledRef.current) return;
-        setState('idle');
+        setItemState('idle');
         setTranscript(null);
       });
     },
     [structure, binName, stableExistingItems, locationId, showToast, clearStructured],
   );
 
-  const handleStopInternal = useCallback(async () => {
-    const handle = recordingRef.current;
-    if (!handle) return;
-    recordingRef.current = null;
+  const handleTranscribed = useCallback((text: string) => {
+    setTranscript(text);
+    setItemState('previewing-transcript');
+    autoAdvanceRef.current = setTimeout(() => {
+      autoAdvanceRef.current = null;
+      if (!cancelledRef.current) advanceToStructuring(text);
+    }, 1500);
+  }, [advanceToStructuring]);
 
-    setState('transcribing');
-    try {
-      const blob = await handle.stop();
-      if (blob.size < 1000) {
-        showToast({ message: 'No audio recorded', variant: 'warning' });
-        setState('idle');
-        return;
-      }
+  const transcription = useTranscription({ onTranscribed: handleTranscribed });
 
-      const formData = new FormData();
-      let ext = 'webm';
-      if (blob.type.includes('mp4')) ext = 'mp4';
-      else if (blob.type.includes('ogg')) ext = 'ogg';
-      formData.append('audio', blob, `recording.${ext}`);
-      const result = await apiFetch<TranscribeResponse>('/api/ai/transcribe', {
-        method: 'POST',
-        body: formData,
-      });
-      notify(Events.PLAN); // AI credit used
+  // Composite state: transcription states pass through, item states overlay
+  const state: DictationState = itemState !== 'idle' ? itemState : transcription.state;
 
-      if (cancelledRef.current) return;
-
-      const text = result.text.trim();
-      if (!text) {
-        showToast({ message: 'No speech detected. Try again.', variant: 'warning' });
-        setState('idle');
-        return;
-      }
-
-      setTranscript(text);
-      setState('previewing-transcript');
-
-      // Auto-advance after 1.5s
-      autoAdvanceRef.current = setTimeout(() => {
-        autoAdvanceRef.current = null;
-        if (!cancelledRef.current) advanceToStructuring(text);
-      }, 1500);
-    } catch (err) {
-      if (cancelledRef.current) return;
-      const message = mapAiError(err, 'Transcription failed. Try again.');
-      setError(message);
-      showToast({
-        message,
-        variant: 'error',
-        action: { label: 'Retry', onClick: () => startRef.current() },
-      });
-      setState('idle');
+  const cancel = useCallback(() => {
+    cancelledRef.current = true;
+    transcription.cancel();
+    if (autoAdvanceRef.current) {
+      clearTimeout(autoAdvanceRef.current);
+      autoAdvanceRef.current = null;
     }
-  }, [showToast, advanceToStructuring]);
-
-  const start = useCallback(async () => {
-    if (state !== 'idle') return;
-    cancelledRef.current = false;
-    setError(null);
+    setItemState('idle');
     setTranscript(null);
     clearStructured();
+  }, [transcription, clearStructured]);
 
-    try {
-      const handle = await startRecording(() => {
-        // Auto-stop callback: proceed to transcription
-        handleStopInternal();
-      });
-      recordingRef.current = handle;
-      setState('recording');
-    } catch (err) {
-      let message = 'Could not start recording';
-      if (err instanceof DOMException) {
-        if (err.name === 'NotAllowedError') message = 'Microphone access was denied by the browser';
-        else if (err.name === 'NotFoundError') message = 'No microphone found on this device';
-        else if (err.name === 'NotReadableError') message = 'Microphone is in use by another application';
-        else message = `Recording error: ${err.name}`;
-      }
-      showToast({ message, variant: 'error' });
-    }
-  }, [state, clearStructured, showToast, handleStopInternal]);
-  startRef.current = start;
-
-  const stop = useCallback(() => {
-    handleStopInternal();
-  }, [handleStopInternal]);
+  const start = useCallback(async () => {
+    cancelledRef.current = false;
+    setTranscript(null);
+    clearStructured();
+    setItemState('idle');
+    transcription.start();
+  }, [transcription, clearStructured]);
 
   const editTranscript = useCallback(
     (text: string) => {
-      // Pause auto-advance
       if (autoAdvanceRef.current) {
         clearTimeout(autoAdvanceRef.current);
         autoAdvanceRef.current = null;
@@ -243,9 +143,8 @@ export function useDictation(options: UseDictationOptions) {
       } catch {
         showToast({ message: 'Failed to add items', variant: 'error' });
       }
-      setState('idle');
+      setItemState('idle');
       setTranscript(null);
-      setDuration(0);
       clearStructured();
     },
     [binId, onAdd, showToast, clearStructured],
@@ -254,11 +153,11 @@ export function useDictation(options: UseDictationOptions) {
   return {
     state,
     transcript,
-    error,
-    duration,
+    error: transcription.error,
+    duration: transcription.duration,
     structuredItems,
     start,
-    stop,
+    stop: transcription.stop,
     cancel,
     confirm,
     editTranscript,
